Move appdata files to new package
