+++ config/config.go
@@ -9,7 +9,6 @@ import (
 	"os"
 	"path/filepath"
 	"runtime"
-	"sync"
 	"time"
 )
 
@@ -23,67 +22,6 @@ var S3Bucket string
 var S3Path string
 var RefreshInterval time.Duration
 var UseJsonLogging bool
-
-var Mu = &sync.Mutex{}
-
-var SupportedJsons = map[string]AppData{
-	"hypermode.json": &HypermodeData,
-	"models.json":    &ModelData,
-}
-
-// HypermodeJson holds the hypermode.json data
-var HypermodeData HypermodeAppData = HypermodeAppData{}
-
-// ModelJson holds the models.json data
-var ModelData ModelsAppData = ModelsAppData{}
-
-// AppData interface
-type AppData any
-
-// struct that holds the hypermode.json data
-type HypermodeAppData struct {
-	Models               []Model               `json:"models"`
-	EmbeddingSpecs       []EmbeddingSpec       `json:"embeddingSpecs"`
-	TrainingInstructions []TrainingInstruction `json:"trainingInstructions"`
-	AppData
-}
-
-type ModelsAppData struct {
-	AppData
-}
-
-type ModelTask string
-
-const (
-	ClassificationTask ModelTask = "classification"
-	EmbeddingTask      ModelTask = "embedding"
-	GeneratorTask      ModelTask = "generator"
-)
-
-type Model struct {
-	Name        string    `json:"name"`
-	Task        ModelTask `json:"task"`
-	SourceModel string    `json:"sourceModel"`
-	Provider    string    `json:"provider"`
-	Host        string    `json:"host"`
-	Endpoint    string    `json:"endpoint"`
-	AuthHeader  string    `json:"authHeader"`
-}
-
-type EmbeddingSpec struct {
-	EntityType string `json:"entityType"`
-	Attribute  string `json:"attribute"`
-	ModelName  string `json:"modelName"`
-	Config     struct {
-		Query    string `json:"query"`
-		Template string `json:"template"`
-	} `json:"config"`
-}
-
-type TrainingInstruction struct {
-	ModelName string   `json:"modelName"`
-	Labels    []string `json:"labels"`
-}
 
 func ParseCommandLineFlags() {
 	flag.IntVar(&Port, "port", 8686, "The HTTP port to listen on.")
